refactor: migrate type of X,Y,vx,vy of flying object to double

+++ src/edu/hitsz/aircraft/AbstractAircraft.java
@@ -19,7 +19,7 @@ public abstract class AbstractAircraft extends AbstractFlyingObject {
     protected int maxHp;
     protected int hp;
 
-    public AbstractAircraft(int locationX, int locationY, int speedX, int speedY, int hp, BaseShootStrategy shootStrategy) {
+    public AbstractAircraft(double locationX, double locationY, double speedX, double speedY, int hp, BaseShootStrategy shootStrategy) {
         super(locationX, locationY, speedX, speedY);
         this.hp = hp;
         this.maxHp = hp;
