+++ src/edu/hitsz/bullet/EnemyBullet.java
@@ -5,7 +5,7 @@ package edu.hitsz.bullet;
  */
 public class EnemyBullet extends BaseBullet {
 
-    public EnemyBullet(int locationX, int locationY, int speedX, int speedY, int power) {
+    public EnemyBullet(double locationX, double locationY, double speedX, double speedY, int power) {
         super(locationX, locationY, speedX, speedY, power);
     }
