+++ src/edu/hitsz/bullet/HeroBullet.java
@@ -5,7 +5,7 @@ package edu.hitsz.bullet;
  */
 public class HeroBullet extends BaseBullet {
 
-    public HeroBullet(int locationX, int locationY, int speedX, int speedY, int power) {
+    public HeroBullet(double locationX, double locationY, double speedX, double speedY, int power) {
         super(locationX, locationY, speedX, speedY, power);
     }
