+++ src/edu/hitsz/supply/BombSupply.java
@@ -3,7 +3,7 @@ package edu.hitsz.supply;
 import edu.hitsz.application.WorldHandle;
 
 public class BombSupply extends BaseSupply{
-    public BombSupply(int x, int y, int speedX, int speedY) {
+    public BombSupply(double x, double y, double speedX, double speedY) {
         super(x, y, speedX, speedY);
     }
